User restriction and Chat room messages added

+++ Community/views.py
@@ -1,7 +1,9 @@
+from django.http import HttpResponse
 from django.shortcuts import render,redirect
 from django.db.models import Q
-from . models import Room,Topic
+from . models import Room,Topic,Message
 from . forms import RoomForm
+from django.contrib.auth.decorators import login_required
 # Create your views here.
 
 
@@ -12,6 +14,7 @@ from . forms import RoomForm
 # ]
 
 
+@login_required(login_url='login')
 def community(request):
     
     q=request.GET.get('q') if request.GET.get('q') != None else ''
@@ -24,16 +27,29 @@ def community(request):
     print(room_count)
     return render(request,"Community/community.html",{'rooms':rooms, 'topics':topics, 'room_count':room_count})
 
-# rooms = Room.objects.all()
 
+@login_required(login_url='')
 def room(request,pk):
     room = Room.objects.get(id=pk)
-    context = {'room':room}
+    room_messages = room.message_set.all().order_by('-created')
+    participants = room.participant.all()
+    
+    if request.method == "POST":
+        message = Message.objects.create(
+            
+            user = request.user,
+            room = room,
+            body = request.POST.get('body')
+            
+        )
+        room.participant.add(request.user)
+        return redirect('room',pk=room.id)
+    context = {'room':room,'room_messages':room_messages,'participants':participants}
     
     return render(request,'Community/room.html',context)
 
 
-
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm
     
@@ -48,11 +64,14 @@ def createRoom(request):
     context = {'form':form}
     return render(request,'Community/room_form.html',context)
 
-
+@login_required(login_url='login')
 def updateRoom(request,pk):
     
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here')
 
     if request.method == "POST":
         form = RoomForm(request.POST,instance=room)
@@ -63,14 +82,29 @@ def updateRoom(request,pk):
     
     return render(request,'Community/room_form.html',context)
 
-
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     
     room = Room.objects.get(id=pk)
-    
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here')
+
     if request.method == "POST":
         room.delete()
         return redirect('community')
     return render(request,'Community/delete.html',{'obj':room})
+
+
+@login_required(login_url='login')
+def deleteMessage(request,pk):
+    
+    message = Message.objects.get(id=pk)
+    if request.user != message.user:
+        return HttpResponse('You are not allowed here')
+
+    if request.method == "POST":
+        message.delete()
+        return redirect('community')
+    return render(request,'Community/delete.html',{'obj':message})
